requirements.txt
# cairosvg==2.7.0
httpx_oauth==0.13.0
Markdown==3.4.3
matplotlib==3.6.3
Pillow==10.0.0
pyshorteners==1.0.1
qrcode==7.4.2
Requests==2.31.0
streamlit==1.22.0
selenium==4.10.0
webdriver-manager